pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "loracamp"
version = "0.1.0"
description = "A Python port of faircamp"
authors = [
    { name = "scruffynerf" }
]
license = { text = "AGPL-3.0" }
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "tomli",
    "tomli-w",
    "jinja2",
    "python-ffmpeg",
    "babel",
    "Pillow",
    "static-ffmpeg",
    "python-ffmpegio"
]

[tool.setuptools]
packages = ["loracamp"]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
pythonpath = ["."]
